+++ src/private.rs
@@ -5,6 +5,7 @@ use crate::types::{
     color::Colour,
     dynamic_type::DynamicType,
     exolvl::Exolvl,
+    function_call::{CallParameter, FunctionCall},
     image::Image,
     layer::Layer,
     level_data::LevelData,
@@ -15,6 +16,7 @@ use crate::types::{
     },
     object::Object,
     object_property::ObjectProperty,
+    old_script::{OldAction, OldActionProperty, OldActionType, Script},
     pattern::Pattern,
     prefab::Prefab,
     varint::Varint,
@@ -60,17 +62,17 @@ impl_sealed!(
     Brush,
     BrushObject,
     BrushGrid,
-    super::Script,
+    Script,
     NovaScript,
-    super::OldAction,
-    super::OldActionType,
-    super::OldActionProperty,
+    OldAction,
+    OldActionType,
+    OldActionProperty,
     Action,
     ActionType,
     NovaValue,
     DynamicType,
-    super::FunctionCall,
-    super::CallParameter,
+    FunctionCall,
+    CallParameter,
     Variable,
     StaticType,
     Activator,
