+++ src/types/action_type.rs
@@ -1,5 +1,8 @@
-use super::novascript::{action::Action, nova_value::NovaValue};
-use crate::{error::Error, FunctionCall, Read, ReadContext, Write};
+use super::{
+    function_call::FunctionCall,
+    novascript::{action::Action, nova_value::NovaValue},
+};
+use crate::{error::Error, Read, ReadContext, Write};
 
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[derive(Clone, Debug)]
