+++ src/types/mod.rs
@@ -4,6 +4,7 @@ pub mod brush;
 pub mod color;
 pub mod dynamic_type;
 pub mod exolvl;
+pub mod function_call;
 pub mod image;
 pub mod layer;
 pub mod level_data;
@@ -11,6 +12,7 @@ pub mod local_level;
 pub mod novascript;
 pub mod object;
 pub mod object_property;
+pub mod old_script;
 pub mod pattern;
 pub mod prefab;
 pub(crate) mod varint;
